requirements.txt
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
Pillow==9.1.1
pycodestyle==2.8.0
requests==2.28.0
soupsieve==2.3.2.post1
toml==0.10.2
urllib3==1.26.9
